# ecommerce-controlled/identity-service/src/main/java/com/ecommerce/identity/web/dto/LoginResponse.java
package com.ecommerce.identity.web.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

/**
 * Login response DTO
 */
@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class LoginResponse {
    
    private String accessToken;
    private String refreshToken;
    private Integer expiresIn;
    private String tokenType;
    private String userId;
}