# ecommerce-controlled/identity-service/src/main/java/com/ecommerce/identity/web/dto/RegisterResponse.java
package com.ecommerce.identity.web.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.util.UUID;

/**
 * Register user response DTO
 */
@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class RegisterResponse {
    
    private UUID userId;
    private String email;
    private String username;
    private String message;
}